refactor(gui): extract app runtime module
