+++ crates/gwt/src/main.rs
@@ -31,10 +31,18 @@ use tokio::runtime::Runtime;
 use uuid::Uuid;
 use wry::WebViewBuilder;
 
+mod app_runtime;
 mod embedded_server;
 mod embedded_web;
 mod repo_browser;
 
+#[cfg(test)]
+pub(crate) use app_runtime::{build_frontend_sync_events, LaunchWizardSession};
+pub(crate) use app_runtime::{
+    ActiveAgentSession, AppEventProxy, AppRuntime, BlockingTaskSpawner, DispatchTarget,
+    IssueLaunchWizardPrepared, OutboundEvent, ProcessLaunch, ProjectOpenTarget, ProjectTabRuntime,
+    WindowAddress,
+};
 #[cfg(test)]
 use embedded_server::{broadcast_runtime_hook_event, health_handler, hook_forward_authorized};
 use embedded_server::{ClientHub, EmbeddedServer};
@@ -110,2499 +118,6 @@ enum UserEvent {
     UpdateAvailable(gwt_core::update::UpdateState),
     #[cfg(target_os = "macos")]
     MenuEvent(muda::MenuEvent),
-}
-
-#[derive(Clone)]
-enum AppEventProxy {
-    Real(EventLoopProxy<UserEvent>),
-    #[cfg(test)]
-    Stub(Arc<Mutex<Vec<UserEvent>>>),
-}
-
-impl AppEventProxy {
-    fn new(proxy: EventLoopProxy<UserEvent>) -> Self {
-        Self::Real(proxy)
-    }
-
-    fn send(&self, event: UserEvent) {
-        match self {
-            Self::Real(proxy) => {
-                let _ = proxy.send_event(event);
-            }
-            #[cfg(test)]
-            Self::Stub(events) => {
-                if let Ok(mut events) = events.lock() {
-                    events.push(event);
-                }
-            }
-        }
-    }
-
-    #[cfg(test)]
-    fn stub() -> (Self, Arc<Mutex<Vec<UserEvent>>>) {
-        let events = Arc::new(Mutex::new(Vec::new()));
-        (Self::Stub(events.clone()), events)
-    }
-}
-
-#[derive(Clone)]
-enum BlockingTaskSpawner {
-    Tokio(tokio::runtime::Handle),
-    #[cfg(test)]
-    Thread,
-}
-
-impl BlockingTaskSpawner {
-    fn tokio(handle: tokio::runtime::Handle) -> Self {
-        Self::Tokio(handle)
-    }
-
-    #[cfg(test)]
-    fn thread() -> Self {
-        Self::Thread
-    }
-
-    fn spawn<F>(&self, task: F)
-    where
-        F: FnOnce() + Send + 'static,
-    {
-        match self {
-            Self::Tokio(handle) => {
-                drop(handle.spawn_blocking(task));
-            }
-            #[cfg(test)]
-            Self::Thread => {
-                thread::Builder::new()
-                    .name("gwt-blocking-task".to_string())
-                    .spawn(task)
-                    .expect("spawn test blocking task");
-            }
-        }
-    }
-}
-
-struct WindowRuntime {
-    pane: Arc<Mutex<Pane>>,
-    /// Handle to the background reader thread that forwards PTY output.
-    /// Taken and joined during `stop_window_runtime` so the reader releases
-    /// its Arc clone of `pane` before the runtime is fully torn down.
-    output_thread: Option<JoinHandle<()>>,
-}
-
-#[derive(Debug, Clone)]
-struct ProcessLaunch {
-    command: String,
-    args: Vec<String>,
-    env: HashMap<String, String>,
-    cwd: Option<PathBuf>,
-}
-
-#[derive(Debug, Clone)]
-struct ActiveAgentSession {
-    window_id: String,
-    session_id: String,
-    agent_id: String,
-    branch_name: String,
-    display_name: String,
-    worktree_path: PathBuf,
-    tab_id: String,
-}
-
-#[derive(Debug, Clone)]
-enum DispatchTarget {
-    Broadcast,
-    Client(ClientId),
-}
-
-#[derive(Debug, Clone)]
-struct OutboundEvent {
-    target: DispatchTarget,
-    event: BackendEvent,
-}
-
-impl OutboundEvent {
-    fn broadcast(event: BackendEvent) -> Self {
-        Self {
-            target: DispatchTarget::Broadcast,
-            event,
-        }
-    }
-
-    fn reply(client_id: impl Into<ClientId>, event: BackendEvent) -> Self {
-        Self {
-            target: DispatchTarget::Client(client_id.into()),
-            event,
-        }
-    }
-}
-
-fn build_frontend_sync_events(
-    client_id: &str,
-    workspace: gwt::AppStateView,
-    terminal_statuses: Vec<(String, WindowProcessStatus, String)>,
-    terminal_snapshots: Vec<(String, Vec<u8>)>,
-    launch_wizard: Option<gwt::LaunchWizardView>,
-    pending_update: Option<gwt_core::update::UpdateState>,
-) -> Vec<OutboundEvent> {
-    let mut events = vec![OutboundEvent::reply(
-        client_id,
-        BackendEvent::WorkspaceState { workspace },
-    )];
-
-    for (id, status, detail) in terminal_statuses {
-        events.push(OutboundEvent::reply(
-            client_id,
-            BackendEvent::TerminalStatus {
-                id,
-                status,
-                detail: Some(detail),
-            },
-        ));
-    }
-
-    for (id, snapshot) in terminal_snapshots {
-        events.push(OutboundEvent::reply(
-            client_id,
-            BackendEvent::TerminalSnapshot {
-                id,
-                data_base64: base64::engine::general_purpose::STANDARD.encode(snapshot),
-            },
-        ));
-    }
-
-    if let Some(wizard) = launch_wizard {
-        events.push(OutboundEvent::reply(
-            client_id,
-            BackendEvent::LaunchWizardState {
-                wizard: Some(Box::new(wizard)),
-            },
-        ));
-    }
-
-    if let Some(state) = pending_update {
-        events.push(OutboundEvent::reply(
-            client_id,
-            BackendEvent::UpdateState(state),
-        ));
-    }
-
-    events
-}
-
-#[derive(Debug, Clone)]
-struct ProjectTabRuntime {
-    id: String,
-    title: String,
-    project_root: PathBuf,
-    kind: gwt::ProjectKind,
-    workspace: WorkspaceState,
-}
-
-#[derive(Debug, Clone)]
-struct WindowAddress {
-    tab_id: String,
-    raw_id: String,
-}
-
-#[derive(Debug, Clone)]
-struct LaunchWizardSession {
-    tab_id: String,
-    wizard_id: String,
-    wizard: LaunchWizardState,
-}
-
-#[derive(Debug, Clone)]
-struct IssueLaunchWizardPrepared {
-    client_id: ClientId,
-    id: String,
-    knowledge_kind: KnowledgeKind,
-    tab_id: String,
-    project_root: PathBuf,
-    issue_number: u64,
-    result: Result<String, String>,
-}
-
-#[derive(Debug, Clone)]
-struct ProjectOpenTarget {
-    project_root: PathBuf,
-    title: String,
-    kind: gwt::ProjectKind,
-}
-
-struct AppRuntime {
-    tabs: Vec<ProjectTabRuntime>,
-    active_tab_id: Option<String>,
-    recent_projects: Vec<gwt::RecentProjectEntry>,
-    runtimes: HashMap<String, WindowRuntime>,
-    window_details: HashMap<String, String>,
-    window_lookup: HashMap<String, WindowAddress>,
-    session_state_path: PathBuf,
-    proxy: AppEventProxy,
-    blocking_tasks: BlockingTaskSpawner,
-    sessions_dir: PathBuf,
-    launch_wizard: Option<LaunchWizardSession>,
-    active_agent_sessions: HashMap<String, ActiveAgentSession>,
-    hook_forward_target: Option<HookForwardTarget>,
-    /// Cached update state so late-connecting WebView clients get the toast.
-    pending_update: Option<gwt_core::update::UpdateState>,
-    /// Shared PTY writer registry published to the WebSocket fast-path.
-    pty_writers: PtyWriterRegistry,
-}
-
-impl ProjectTabRuntime {
-    fn from_persisted(
-        tab: gwt::PersistedSessionTabState,
-        workspace: gwt::PersistedWorkspaceState,
-    ) -> Self {
-        Self {
-            id: tab.id,
-            title: tab.title,
-            project_root: tab.project_root,
-            kind: tab.kind,
-            workspace: WorkspaceState::from_persisted(workspace),
-        }
-    }
-}
-
-impl AppRuntime {
-    fn new(
-        proxy: EventLoopProxy<UserEvent>,
-        pty_writers: PtyWriterRegistry,
-        blocking_tasks: BlockingTaskSpawner,
-    ) -> std::io::Result<Self> {
-        let session_state_path = gwt_core::paths::gwt_session_state_path();
-        let launch_dir = std::env::current_dir().unwrap_or_else(|_| PathBuf::from("."));
-        let legacy_target = resolve_project_target(&launch_dir)
-            .unwrap_or_else(|_| fallback_project_target(launch_dir.clone()));
-        migrate_legacy_workspace_state(
-            &gwt::legacy_workspace_state_path(),
-            &session_state_path,
-            &legacy_target.project_root,
-            legacy_target.kind,
-        )?;
-        let persisted = load_session_state(&session_state_path)?;
-        let tabs = persisted
-            .tabs
-            .into_iter()
-            .map(|tab| {
-                let workspace = load_restored_workspace_state(&tab.project_root)?;
-                Ok(ProjectTabRuntime::from_persisted(tab, workspace))
-            })
-            .collect::<std::io::Result<Vec<_>>>()?;
-        let active_tab_id = normalize_active_tab_id(&tabs, persisted.active_tab_id);
-        let sessions_dir = gwt_core::paths::gwt_sessions_dir();
-        let _ = gwt_agent::reset_runtime_state_dir(&sessions_dir);
-
-        let mut app = Self {
-            tabs,
-            active_tab_id,
-            recent_projects: dedupe_recent_projects(persisted.recent_projects),
-            runtimes: HashMap::new(),
-            window_details: HashMap::new(),
-            window_lookup: HashMap::new(),
-            session_state_path,
-            proxy: AppEventProxy::new(proxy),
-            blocking_tasks,
-            sessions_dir,
-            launch_wizard: None,
-            active_agent_sessions: HashMap::new(),
-            hook_forward_target: None,
-            pending_update: None,
-            pty_writers,
-        };
-        app.rebuild_window_lookup();
-        app.seed_restored_window_details();
-        Ok(app)
-    }
-
-    fn bootstrap(&mut self) {
-        let windows = self
-            .tabs
-            .iter()
-            .flat_map(|tab| {
-                tab.workspace
-                    .persisted()
-                    .windows
-                    .clone()
-                    .into_iter()
-                    .map(|window| (tab.id.clone(), window))
-                    .collect::<Vec<_>>()
-            })
-            .collect::<Vec<_>>();
-        for (tab_id, window) in windows {
-            if !should_auto_start_restored_window(&window) {
-                continue;
-            }
-            let _ = self.start_window(&tab_id, &window.id, window.preset, window.geometry.clone());
-        }
-        let _ = self.persist();
-    }
-
-    fn set_hook_forward_target(&mut self, target: HookForwardTarget) {
-        self.hook_forward_target = Some(target);
-    }
-
-    fn handle_frontend_event(
-        &mut self,
-        client_id: ClientId,
-        event: FrontendEvent,
-    ) -> Vec<OutboundEvent> {
-        match event {
-            FrontendEvent::FrontendReady => self.frontend_sync_events(&client_id),
-            FrontendEvent::OpenProjectDialog => self.open_project_dialog_events(),
-            FrontendEvent::ReopenRecentProject { path } => {
-                self.open_project_path_events(PathBuf::from(path))
-            }
-            FrontendEvent::SelectProjectTab { tab_id } => self.select_project_tab_events(&tab_id),
-            FrontendEvent::CloseProjectTab { tab_id } => self.close_project_tab_events(&tab_id),
-            FrontendEvent::CreateWindow { preset, bounds } => {
-                self.create_window_events(preset, bounds)
-            }
-            FrontendEvent::FocusWindow { id, bounds } => self.focus_window_events(&id, bounds),
-            FrontendEvent::CycleFocus { direction, bounds } => {
-                self.cycle_focus_events(direction, bounds)
-            }
-            FrontendEvent::UpdateViewport { viewport } => self.update_viewport_events(viewport),
-            FrontendEvent::ArrangeWindows { mode, bounds } => {
-                self.arrange_windows_events(mode, bounds)
-            }
-            FrontendEvent::MaximizeWindow { id, bounds } => {
-                self.maximize_window_events(&id, bounds)
-            }
-            FrontendEvent::MinimizeWindow { id } => self.minimize_window_events(&id),
-            FrontendEvent::RestoreWindow { id } => self.restore_window_events(&id),
-            FrontendEvent::ListWindows => {
-                vec![OutboundEvent::reply(client_id, self.list_windows_event())]
-            }
-            FrontendEvent::UpdateWindowGeometry {
-                id,
-                geometry,
-                cols,
-                rows,
-            } => self.update_window_geometry_events(&id, geometry, cols, rows),
-            FrontendEvent::CloseWindow { id } => self.close_window_events(&id),
-            FrontendEvent::TerminalInput { id, data } => self.terminal_input_events(&id, &data),
-            FrontendEvent::LoadFileTree { id, path } => {
-                let path = path.unwrap_or_default();
-                vec![OutboundEvent::reply(
-                    client_id,
-                    self.load_file_tree_event(&id, &path),
-                )]
-            }
-            FrontendEvent::LoadBranches { id } => self.load_branches_events(&client_id, &id),
-            FrontendEvent::LoadKnowledgeBridge {
-                id,
-                knowledge_kind,
-                selected_number,
-                refresh,
-            } => self.load_knowledge_bridge_events(
-                &client_id,
-                &id,
-                knowledge_kind,
-                selected_number,
-                refresh,
-            ),
-            FrontendEvent::SelectKnowledgeBridgeEntry {
-                id,
-                knowledge_kind,
-                number,
-            } => self.load_knowledge_bridge_events(
-                &client_id,
-                &id,
-                knowledge_kind,
-                Some(number),
-                false,
-            ),
-            FrontendEvent::RunBranchCleanup {
-                id,
-                branches,
-                delete_remote,
-            } => self.run_branch_cleanup_events(&client_id, &id, &branches, delete_remote),
-            FrontendEvent::OpenIssueLaunchWizard { id, issue_number } => {
-                self.open_issue_launch_wizard_events(&client_id, &id, issue_number)
-            }
-            FrontendEvent::OpenLaunchWizard {
-                id,
-                branch_name,
-                linked_issue_number,
-            } => self.open_launch_wizard(&id, &branch_name, linked_issue_number),
-            FrontendEvent::LaunchWizardAction { action, bounds } => {
-                self.handle_launch_wizard_action(action, bounds)
-            }
-            FrontendEvent::ApplyUpdate => {
-                std::thread::spawn(apply_update_and_exit);
-                vec![]
-            }
-            FrontendEvent::ListCustomAgents => vec![OutboundEvent::reply(
-                client_id,
-                gwt::custom_agents_dispatch::list_event(),
-            )],
-            FrontendEvent::ListCustomAgentPresets => vec![OutboundEvent::reply(
-                client_id,
-                gwt::custom_agents_dispatch::list_presets_event(),
-            )],
-            FrontendEvent::AddCustomAgentFromPreset { input } => vec![OutboundEvent::reply(
-                client_id,
-                gwt::custom_agents_dispatch::add_from_preset_event(input),
-            )],
-            FrontendEvent::UpdateCustomAgent { agent } => vec![OutboundEvent::reply(
-                client_id,
-                gwt::custom_agents_dispatch::update_event(*agent),
-            )],
-            FrontendEvent::DeleteCustomAgent { agent_id } => vec![OutboundEvent::reply(
-                client_id,
-                gwt::custom_agents_dispatch::delete_event(agent_id),
-            )],
-            FrontendEvent::TestBackendConnection { base_url, api_key } => {
-                self.spawn_backend_connection_probe(client_id, base_url, api_key);
-                Vec::new()
-            }
-        }
-    }
-
-    fn spawn_backend_connection_probe(
-        &self,
-        client_id: ClientId,
-        base_url: String,
-        api_key: String,
-    ) {
-        let proxy = self.proxy.clone();
-        self.blocking_tasks.spawn(move || {
-            let event = gwt::custom_agents_dispatch::test_connection_event(&base_url, &api_key);
-            proxy.send(UserEvent::Dispatch(vec![OutboundEvent::reply(
-                client_id, event,
-            )]));
-        });
-    }
-
-    fn frontend_sync_events(&self, client_id: &str) -> Vec<OutboundEvent> {
-        let terminal_statuses = self
-            .window_details
-            .iter()
-            .filter_map(|(id, detail)| {
-                self.window_status(id)
-                    .map(|status| (id.clone(), status, detail.clone()))
-            })
-            .collect();
-        let terminal_snapshots = self
-            .runtimes
-            .iter()
-            .filter_map(|(id, runtime)| {
-                let snapshot = runtime
-                    .pane
-                    .lock()
-                    .map(|pane| pane.screen().contents().into_bytes())
-                    .unwrap_or_default();
-                (!snapshot.is_empty()).then_some((id.clone(), snapshot))
-            })
-            .collect();
-
-        build_frontend_sync_events(
-            client_id,
-            self.app_state_view(),
-            terminal_statuses,
-            terminal_snapshots,
-            self.launch_wizard
-                .as_ref()
-                .map(|wizard| wizard.wizard.view()),
-            self.pending_update.clone(),
-        )
-    }
-
-    fn open_project_dialog_events(&mut self) -> Vec<OutboundEvent> {
-        let selected = rfd::FileDialog::new().pick_folder();
-        let Some(path) = selected else {
-            return Vec::new();
-        };
-        self.open_project_path_events(path)
-    }
-
-    fn open_project_path_events(&mut self, path: PathBuf) -> Vec<OutboundEvent> {
-        match self.open_project_path(path) {
-            Ok(wizard_closed) => {
-                let mut events = vec![self.workspace_state_broadcast()];
-                if wizard_closed {
-                    events.push(self.launch_wizard_state_broadcast(None));
-                }
-                events
-            }
-            Err(error) => vec![OutboundEvent::broadcast(BackendEvent::ProjectOpenError {
-                message: error,
-            })],
-        }
-    }
-
-    fn open_project_path(&mut self, path: PathBuf) -> Result<bool, String> {
-        let target = resolve_project_target(&path)?;
-        if let Some(existing) = self
-            .tabs
-            .iter()
-            .find(|tab| same_worktree_path(&tab.project_root, &target.project_root))
-            .map(|tab| tab.id.clone())
-        {
-            let wizard_closed = self.set_active_tab(existing);
-            self.remember_recent_project(&target);
-            self.persist().map_err(|error| error.to_string())?;
-            return Ok(wizard_closed);
-        }
-
-        let tab_id = format!("project-{}", Uuid::new_v4().simple());
-        self.tabs.push(ProjectTabRuntime {
-            id: tab_id.clone(),
-            title: target.title.clone(),
-            project_root: target.project_root.clone(),
-            kind: target.kind,
-            workspace: WorkspaceState::from_persisted({
-                load_restored_workspace_state(&target.project_root)
-                    .map_err(|error| error.to_string())?
-            }),
-        });
-        self.active_tab_id = Some(tab_id);
-        self.remember_recent_project(&target);
-        let wizard_closed = self.clear_launch_wizard().is_some();
-        self.persist().map_err(|error| error.to_string())?;
-        Ok(wizard_closed)
-    }
-
-    fn remember_recent_project(&mut self, target: &ProjectOpenTarget) {
-        self.recent_projects
-            .retain(|entry| !same_worktree_path(&entry.path, &target.project_root));
-        self.recent_projects.insert(
-            0,
-            gwt::RecentProjectEntry {
-                path: target.project_root.clone(),
-                title: target.title.clone(),
-                kind: target.kind,
-            },
-        );
-        if self.recent_projects.len() > 12 {
-            self.recent_projects.truncate(12);
-        }
-    }
-
-    fn select_project_tab_events(&mut self, tab_id: &str) -> Vec<OutboundEvent> {
-        if !self.tabs.iter().any(|tab| tab.id == tab_id) {
-            return Vec::new();
-        }
-        let wizard_closed = self.set_active_tab(tab_id.to_string());
-        let _ = self.persist();
-        let mut events = vec![self.workspace_state_broadcast()];
-        if wizard_closed {
-            events.push(self.launch_wizard_state_broadcast(None));
-        }
-        events
-    }
-
-    fn close_project_tab_events(&mut self, tab_id: &str) -> Vec<OutboundEvent> {
-        let Some(index) = self.tabs.iter().position(|tab| tab.id == tab_id) else {
-            return Vec::new();
-        };
-
-        let window_ids = self
-            .tabs
-            .get(index)
-            .map(|tab| {
-                tab.workspace
-                    .persisted()
-                    .windows
-                    .iter()
-                    .map(|window| combined_window_id(&tab.id, &window.id))
-                    .collect::<Vec<_>>()
-            })
-            .unwrap_or_default();
-        for window_id in window_ids {
-            self.stop_window_runtime(&window_id);
-            self.window_lookup.remove(&window_id);
-        }
-
-        self.tabs.remove(index);
-        if self.tabs.is_empty() {
-            self.active_tab_id = None;
-        } else if self.active_tab_id.as_deref() == Some(tab_id) {
-            let next_index = index.saturating_sub(1).min(self.tabs.len() - 1);
-            self.active_tab_id = self.tabs.get(next_index).map(|tab| tab.id.clone());
-        }
-
-        let wizard_closed = self
-            .launch_wizard
-            .as_ref()
-            .is_some_and(|wizard| wizard.tab_id == tab_id);
-        if wizard_closed {
-            self.launch_wizard = None;
-        }
-        let _ = self.persist();
-
-        let mut events = vec![self.workspace_state_broadcast()];
-        if wizard_closed {
-            events.push(self.launch_wizard_state_broadcast(None));
-        }
-        events
-    }
-
-    fn create_window_events(
-        &mut self,
-        preset: WindowPreset,
-        bounds: WindowGeometry,
-    ) -> Vec<OutboundEvent> {
-        let Some(tab_id) = self.active_tab_id.clone() else {
-            return Vec::new();
-        };
-        let window = {
-            let Some(tab) = self.tab_mut(&tab_id) else {
-                return Vec::new();
-            };
-            tab.workspace.add_window(preset, bounds)
-        };
-        self.register_window(&tab_id, &window.id);
-        let runtime_event = self.start_window(&tab_id, &window.id, window.preset, window.geometry);
-        let _ = self.persist();
-        let mut events = vec![self.workspace_state_broadcast()];
-        if let Some(event) = runtime_event {
-            events.push(OutboundEvent::broadcast(event));
-        }
-        events
-    }
-
-    fn focus_window_events(
-        &mut self,
-        id: &str,
-        bounds: Option<WindowGeometry>,
-    ) -> Vec<OutboundEvent> {
-        let Some(address) = self.window_lookup.get(id).cloned() else {
-            return Vec::new();
-        };
-        let Some(tab) = self.tab_mut(&address.tab_id) else {
-            return Vec::new();
-        };
-        if !tab.workspace.focus_window(&address.raw_id, bounds) {
-            return Vec::new();
-        }
-        self.active_tab_id = Some(address.tab_id);
-        let _ = self.persist();
-        vec![self.workspace_state_broadcast()]
-    }
-
-    fn cycle_focus_events(
-        &mut self,
-        direction: gwt::FocusCycleDirection,
-        bounds: WindowGeometry,
-    ) -> Vec<OutboundEvent> {
-        let Some(tab) = self.active_tab_mut() else {
-            return Vec::new();
-        };
-        if tab.workspace.cycle_focus(direction, bounds).is_none() {
-            return Vec::new();
-        }
-        let _ = self.persist();
-        vec![self.workspace_state_broadcast()]
-    }
-
-    fn update_viewport_events(&mut self, viewport: gwt::CanvasViewport) -> Vec<OutboundEvent> {
-        let Some(tab) = self.active_tab_mut() else {
-            return Vec::new();
-        };
-        tab.workspace.update_viewport(viewport);
-        let _ = self.persist();
-        vec![self.workspace_state_broadcast()]
-    }
-
-    fn arrange_windows_events(
-        &mut self,
-        mode: gwt::ArrangeMode,
-        bounds: WindowGeometry,
-    ) -> Vec<OutboundEvent> {
-        let Some(tab_id) = self.active_tab_id.clone() else {
-            return Vec::new();
-        };
-        let arranged = {
-            let Some(tab) = self.tab_mut(&tab_id) else {
-                return Vec::new();
-            };
-            tab.workspace.arrange_windows(mode, bounds)
-        };
-        if !arranged {
-            return Vec::new();
-        }
-        if let Some(tab) = self.tab(&tab_id) {
-            for window in tab.workspace.persisted().windows.iter() {
-                self.resize_runtime_to_window(&combined_window_id(&tab_id, &window.id));
-            }
-        }
-        let _ = self.persist();
-        vec![self.workspace_state_broadcast()]
-    }
-
-    fn maximize_window_events(&mut self, id: &str, bounds: WindowGeometry) -> Vec<OutboundEvent> {
-        let Some(address) = self.window_lookup.get(id).cloned() else {
-            return Vec::new();
-        };
-        let updated = {
-            let Some(tab) = self.tab_mut(&address.tab_id) else {
-                return Vec::new();
-            };
-            tab.workspace.maximize_window(&address.raw_id, bounds)
-        };
-        if !updated {
-            return Vec::new();
-        }
-        let _ = self.set_active_tab(address.tab_id);
-        self.resize_runtime_to_window(id);
-        let _ = self.persist();
-        vec![self.workspace_state_broadcast()]
-    }
-
-    fn minimize_window_events(&mut self, id: &str) -> Vec<OutboundEvent> {
-        let Some(address) = self.window_lookup.get(id).cloned() else {
-            return Vec::new();
-        };
-        let updated = {
-            let Some(tab) = self.tab_mut(&address.tab_id) else {
-                return Vec::new();
-            };
-            tab.workspace.minimize_window(&address.raw_id)
-        };
-        if !updated {
-            return Vec::new();
-        }
-        let _ = self.set_active_tab(address.tab_id);
-        self.resize_runtime_to_window(id);
-        let _ = self.persist();
-        vec![self.workspace_state_broadcast()]
-    }
-
-    fn restore_window_events(&mut self, id: &str) -> Vec<OutboundEvent> {
-        let Some(address) = self.window_lookup.get(id).cloned() else {
-            return Vec::new();
-        };
-        let updated = {
-            let Some(tab) = self.tab_mut(&address.tab_id) else {
-                return Vec::new();
-            };
-            tab.workspace.restore_window(&address.raw_id)
-        };
-        if !updated {
-            return Vec::new();
-        }
-        let _ = self.set_active_tab(address.tab_id);
-        self.resize_runtime_to_window(id);
-        let _ = self.persist();
-        vec![self.workspace_state_broadcast()]
-    }
-
-    fn update_window_geometry_events(
-        &mut self,
-        id: &str,
-        geometry: WindowGeometry,
-        cols: u16,
-        rows: u16,
-    ) -> Vec<OutboundEvent> {
-        let Some(address) = self.window_lookup.get(id).cloned() else {
-            return Vec::new();
-        };
-        let updated = {
-            let Some(tab) = self.tab_mut(&address.tab_id) else {
-                return Vec::new();
-            };
-            tab.workspace.update_geometry(&address.raw_id, geometry)
-        };
-        if !updated {
-            return Vec::new();
-        }
-        if let Some(runtime) = self.runtimes.get(id) {
-            if let Ok(mut pane) = runtime.pane.lock() {
-                let _ = pane.resize(cols.max(20), rows.max(6));
-            }
-        }
-        let _ = self.persist();
-        vec![self.workspace_state_broadcast()]
-    }
-
-    fn close_window_events(&mut self, id: &str) -> Vec<OutboundEvent> {
-        self.stop_window_runtime(id);
-        if !close_window_from_workspace(
-            &mut self.tabs,
-            &mut self.window_lookup,
-            &mut self.window_details,
-            id,
-        ) {
-            return Vec::new();
-        }
-        let _ = self.persist();
-        vec![self.workspace_state_broadcast()]
-    }
-
-    fn list_windows_event(&self) -> BackendEvent {
-        let windows = self
-            .active_tab_id
-            .as_ref()
-            .and_then(|tab_id| self.tab(tab_id))
-            .map(|tab| workspace_view_for_tab(tab).windows)
-            .unwrap_or_default();
-        BackendEvent::WindowList { windows }
-    }
-
-    fn terminal_input_events(&mut self, id: &str, data: &str) -> Vec<OutboundEvent> {
-        let data_len = data.len();
-        let write_result = {
-            let Some(runtime) = self.runtimes.get(id) else {
-                tracing::debug!(
-                    target: "gwt_input_trace",
-                    stage = "event_loop_runtime_missing",
-                    window_id = %id,
-                    data_len,
-                    "terminal_input dropped: no runtime for window"
-                );
-                return Vec::new();
-            };
-
-            let lock_started = Instant::now();
-            let lock_result = runtime.pane.lock().map_err(|error| error.to_string());
-            let lock_wait_us = lock_started.elapsed().as_micros() as u64;
-
-            match lock_result {
-                Ok(pane) => {
-                    let write_started = Instant::now();
-                    let result = pane
-                        .write_input(data.as_bytes())
-                        .map_err(|error| error.to_string());
-                    tracing::debug!(
-                        target: "gwt_input_trace",
-                        stage = "pty_write",
-                        window_id = %id,
-                        data_len,
-                        lock_wait_us,
-                        write_us = write_started.elapsed().as_micros() as u64,
-                        ok = result.is_ok(),
-                        "terminal_input forwarded to PTY writer"
-                    );
-                    result
-                }
-                Err(error) => {
-                    tracing::debug!(
-                        target: "gwt_input_trace",
-                        stage = "pane_lock_failed",
-                        window_id = %id,
-                        data_len,
-                        lock_wait_us,
-                        error = %error,
-                        "terminal_input dropped: pane mutex poisoned"
-                    );
-                    Err(error)
-                }
-            }
-        };
-
-        match write_result {
-            Ok(()) => Vec::new(),
-            Err(error) => {
-                self.handle_runtime_status(id.to_string(), WindowProcessStatus::Error, Some(error))
-            }
-        }
-    }
-
-    fn load_file_tree_event(&self, id: &str, path: &str) -> BackendEvent {
-        let Some(address) = self.window_lookup.get(id) else {
-            return BackendEvent::FileTreeError {
-                id: id.to_string(),
-                path: path.to_string(),
-                message: "Window not found".to_string(),
-            };
-        };
-        let Some(tab) = self.tab(&address.tab_id) else {
-            return BackendEvent::FileTreeError {
-                id: id.to_string(),
-                path: path.to_string(),
-                message: "Project tab not found".to_string(),
-            };
-        };
-        let Some(window) = tab.workspace.window(&address.raw_id) else {
-            return BackendEvent::FileTreeError {
-                id: id.to_string(),
-                path: path.to_string(),
-                message: "Window not found".to_string(),
-            };
-        };
-
-        if window.preset != WindowPreset::FileTree {
-            return BackendEvent::FileTreeError {
-                id: id.to_string(),
-                path: path.to_string(),
-                message: "Window is not a file tree".to_string(),
-            };
-        }
-
-        let relative_path = if path.is_empty() {
-            None
-        } else {
-            Some(Path::new(path))
-        };
-
-        match list_directory_entries(&tab.project_root, relative_path) {
-            Ok(entries) => BackendEvent::FileTreeEntries {
-                id: id.to_string(),
-                path: path.to_string(),
-                entries,
-            },
-            Err(error) => BackendEvent::FileTreeError {
-                id: id.to_string(),
-                path: path.to_string(),
-                message: error.to_string(),
-            },
-        }
-    }
-
-    fn load_branches_events(&self, client_id: &str, id: &str) -> Vec<OutboundEvent> {
-        let Some(address) = self.window_lookup.get(id) else {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::BranchError {
-                    id: id.to_string(),
-                    message: "Window not found".to_string(),
-                },
-            )];
-        };
-        let Some(tab) = self.tab(&address.tab_id) else {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::BranchError {
-                    id: id.to_string(),
-                    message: "Project tab not found".to_string(),
-                },
-            )];
-        };
-        let Some(window) = tab.workspace.window(&address.raw_id) else {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::BranchError {
-                    id: id.to_string(),
-                    message: "Window not found".to_string(),
-                },
-            )];
-        };
-
-        if window.preset != WindowPreset::Branches {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::BranchError {
-                    id: id.to_string(),
-                    message: "Window is not a branches list".to_string(),
-                },
-            )];
-        }
-
-        spawn_branch_load_async(
-            self.proxy.clone(),
-            client_id.to_string(),
-            id.to_string(),
-            tab.project_root.clone(),
-            self.active_session_branches_for_tab(&address.tab_id),
-        );
-        Vec::new()
-    }
-
-    fn load_knowledge_bridge_events(
-        &self,
-        client_id: &str,
-        id: &str,
-        kind: KnowledgeKind,
-        selected_number: Option<u64>,
-        refresh: bool,
-    ) -> Vec<OutboundEvent> {
-        let Some(address) = self.window_lookup.get(id) else {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::KnowledgeError {
-                    id: id.to_string(),
-                    knowledge_kind: kind,
-                    message: "Window not found".to_string(),
-                },
-            )];
-        };
-        let Some(tab) = self.tab(&address.tab_id) else {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::KnowledgeError {
-                    id: id.to_string(),
-                    knowledge_kind: kind,
-                    message: "Project tab not found".to_string(),
-                },
-            )];
-        };
-        let Some(window) = tab.workspace.window(&address.raw_id) else {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::KnowledgeError {
-                    id: id.to_string(),
-                    knowledge_kind: kind,
-                    message: "Window not found".to_string(),
-                },
-            )];
-        };
-        if knowledge_kind_for_preset(window.preset) != Some(kind) {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::KnowledgeError {
-                    id: id.to_string(),
-                    knowledge_kind: kind,
-                    message: "Window is not a knowledge bridge".to_string(),
-                },
-            )];
-        }
-
-        match load_knowledge_bridge(&tab.project_root, kind, selected_number, refresh) {
-            Ok(view) => vec![
-                OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::KnowledgeEntries {
-                        id: id.to_string(),
-                        knowledge_kind: kind,
-                        entries: view.entries,
-                        selected_number: view.selected_number,
-                        empty_message: view.empty_message,
-                        refresh_enabled: view.refresh_enabled,
-                    },
-                ),
-                OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::KnowledgeDetail {
-                        id: id.to_string(),
-                        knowledge_kind: kind,
-                        detail: view.detail,
-                    },
-                ),
-            ],
-            Err(error) => vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::KnowledgeError {
-                    id: id.to_string(),
-                    knowledge_kind: kind,
-                    message: error,
-                },
-            )],
-        }
-    }
-
-    fn run_branch_cleanup_events(
-        &self,
-        client_id: &str,
-        id: &str,
-        branches: &[String],
-        delete_remote: bool,
-    ) -> Vec<OutboundEvent> {
-        let Some(address) = self.window_lookup.get(id) else {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::BranchError {
-                    id: id.to_string(),
-                    message: "Window not found".to_string(),
-                },
-            )];
-        };
-        let Some(tab) = self.tab(&address.tab_id) else {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::BranchError {
-                    id: id.to_string(),
-                    message: "Project tab not found".to_string(),
-                },
-            )];
-        };
-        let Some(window) = tab.workspace.window(&address.raw_id) else {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::BranchError {
-                    id: id.to_string(),
-                    message: "Window not found".to_string(),
-                },
-            )];
-        };
-
-        if window.preset != WindowPreset::Branches {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::BranchError {
-                    id: id.to_string(),
-                    message: "Window is not a branches list".to_string(),
-                },
-            )];
-        }
-
-        spawn_branch_cleanup_async(
-            self.proxy.clone(),
-            client_id.to_string(),
-            id.to_string(),
-            tab.project_root.clone(),
-            self.active_session_branches_for_tab(&address.tab_id),
-            branches.to_vec(),
-            delete_remote,
-        );
-        Vec::new()
-    }
-}
-
-fn spawn_branch_cleanup_async(
-    proxy: AppEventProxy,
-    client_id: ClientId,
-    window_id: String,
-    project_root: PathBuf,
-    active_session_branches: std::collections::HashSet<String>,
-    branches: Vec<String>,
-    delete_remote: bool,
-) {
-    thread::spawn(move || {
-        let events =
-            match list_branch_entries_with_active_sessions(&project_root, &active_session_branches)
-            {
-                Ok(entries) => {
-                    let results = cleanup_selected_branches(
-                        &project_root,
-                        &entries,
-                        &branches,
-                        delete_remote,
-                    );
-                    let mut events = vec![OutboundEvent::reply(
-                        client_id.clone(),
-                        BackendEvent::BranchCleanupResult {
-                            id: window_id.clone(),
-                            results,
-                        },
-                    )];
-                    match list_branch_entries_with_active_sessions(
-                        &project_root,
-                        &active_session_branches,
-                    ) {
-                        Ok(entries) => events.push(OutboundEvent::reply(
-                            client_id.clone(),
-                            BackendEvent::BranchEntries {
-                                id: window_id.clone(),
-                                phase: BranchEntriesPhase::Hydrated,
-                                entries,
-                            },
-                        )),
-                        Err(error) => events.push(OutboundEvent::reply(
-                            client_id.clone(),
-                            BackendEvent::BranchError {
-                                id: window_id.clone(),
-                                message: error.to_string(),
-                            },
-                        )),
-                    }
-                    events
-                }
-                Err(error) => vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::BranchError {
-                        id: window_id,
-                        message: error.to_string(),
-                    },
-                )],
-            };
-        proxy.send(UserEvent::Dispatch(events));
-    });
-}
-
-impl AppRuntime {
-    fn open_launch_wizard(
-        &mut self,
-        id: &str,
-        branch_name: &str,
-        linked_issue_number: Option<u64>,
-    ) -> Vec<OutboundEvent> {
-        let Some(address) = self.window_lookup.get(id).cloned() else {
-            return vec![OutboundEvent::broadcast(BackendEvent::BranchError {
-                id: id.to_string(),
-                message: "Window not found".to_string(),
-            })];
-        };
-        let Some(tab) = self.tab(&address.tab_id) else {
-            return vec![OutboundEvent::broadcast(BackendEvent::BranchError {
-                id: id.to_string(),
-                message: "Project tab not found".to_string(),
-            })];
-        };
-        let Some(window) = tab.workspace.window(&address.raw_id) else {
-            return vec![OutboundEvent::broadcast(BackendEvent::BranchError {
-                id: id.to_string(),
-                message: "Window not found".to_string(),
-            })];
-        };
-
-        if window.preset != WindowPreset::Branches {
-            return vec![OutboundEvent::broadcast(BackendEvent::BranchError {
-                id: id.to_string(),
-                message: "Window is not a branches list".to_string(),
-            })];
-        }
-
-        let project_root = tab.project_root.clone();
-        let tab_id = address.tab_id.clone();
-        match self.open_launch_wizard_for_branch(
-            &tab_id,
-            &project_root,
-            branch_name,
-            linked_issue_number,
-        ) {
-            Ok(()) => vec![self.launch_wizard_state_outbound()],
-            Err(error) => vec![OutboundEvent::broadcast(BackendEvent::BranchError {
-                id: id.to_string(),
-                message: error,
-            })],
-        }
-    }
-
-    fn open_launch_wizard_for_branch(
-        &mut self,
-        tab_id: &str,
-        project_root: &Path,
-        branch_name: &str,
-        linked_issue_number: Option<u64>,
-    ) -> Result<(), String> {
-        let normalized_branch_name = normalize_branch_name(branch_name);
-        let live_sessions = self.live_sessions_for_branch(tab_id, &normalized_branch_name);
-        let wizard_id = Uuid::new_v4().to_string();
-        self.launch_wizard = Some(LaunchWizardSession {
-            tab_id: tab_id.to_string(),
-            wizard_id: wizard_id.clone(),
-            wizard: LaunchWizardState::open_loading(
-                LaunchWizardContext {
-                    selected_branch: synthetic_branch_entry(branch_name),
-                    normalized_branch_name,
-                    worktree_path: None,
-                    quick_start_root: project_root.to_path_buf(),
-                    live_sessions,
-                    docker_context: None,
-                    docker_service_status: gwt_docker::ComposeServiceStatus::NotFound,
-                    linked_issue_number,
-                },
-                Vec::new(),
-            ),
-        });
-
-        let proxy = self.proxy.clone();
-        let sessions_dir = self.sessions_dir.clone();
-        let project_root = project_root.to_path_buf();
-        let branch_name = branch_name.to_string();
-        let active_session_branches = self.active_session_branches_for_tab(tab_id);
-        thread::spawn(move || {
-            let result = resolve_launch_wizard_hydration(
-                &project_root,
-                &branch_name,
-                &active_session_branches,
-                &sessions_dir,
-            );
-            proxy.send(UserEvent::LaunchWizardHydrated { wizard_id, result });
-        });
-
-        Ok(())
-    }
-
-    fn open_issue_launch_wizard_events(
-        &mut self,
-        client_id: &str,
-        id: &str,
-        issue_number: u64,
-    ) -> Vec<OutboundEvent> {
-        let Some(address) = self.window_lookup.get(id).cloned() else {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::KnowledgeError {
-                    id: id.to_string(),
-                    knowledge_kind: KnowledgeKind::Issue,
-                    message: "Window not found".to_string(),
-                },
-            )];
-        };
-        let Some(tab) = self.tab(&address.tab_id) else {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::KnowledgeError {
-                    id: id.to_string(),
-                    knowledge_kind: KnowledgeKind::Issue,
-                    message: "Project tab not found".to_string(),
-                },
-            )];
-        };
-        let Some(window) = tab.workspace.window(&address.raw_id) else {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::KnowledgeError {
-                    id: id.to_string(),
-                    knowledge_kind: KnowledgeKind::Issue,
-                    message: "Window not found".to_string(),
-                },
-            )];
-        };
-        let Some(kind) = knowledge_kind_for_preset(window.preset) else {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::KnowledgeError {
-                    id: id.to_string(),
-                    knowledge_kind: KnowledgeKind::Issue,
-                    message: "Window is not a knowledge bridge".to_string(),
-                },
-            )];
-        };
-
-        let project_root = tab.project_root.clone();
-        let tab_id = address.tab_id.clone();
-        let proxy = self.proxy.clone();
-        let client_id = client_id.to_string();
-        let id = id.to_string();
-        let active_session_branches = self.active_session_branches_for_tab(&address.tab_id);
-        thread::spawn(move || {
-            let result =
-                list_branch_entries_with_active_sessions(&project_root, &active_session_branches)
-                    .map_err(|error| error.to_string())
-                    .and_then(|entries| {
-                        preferred_issue_launch_branch(&entries)
-                            .ok_or_else(|| "No local branch is available for launch".to_string())
-                    });
-            proxy.send(UserEvent::IssueLaunchWizardPrepared(
-                IssueLaunchWizardPrepared {
-                    client_id,
-                    id,
-                    knowledge_kind: kind,
-                    tab_id,
-                    project_root,
-                    issue_number,
-                    result,
-                },
-            ));
-        });
-        Vec::new()
-    }
-
-    fn handle_launch_wizard_hydrated(
-        &mut self,
-        wizard_id: String,
-        result: Result<LaunchWizardHydration, String>,
-    ) -> Vec<OutboundEvent> {
-        let Some(session) = self.launch_wizard.as_mut() else {
-            return Vec::new();
-        };
-        if session.wizard_id != wizard_id {
-            return Vec::new();
-        }
-
-        match result {
-            Ok(hydration) => session.wizard.apply_hydration(hydration),
-            Err(error) => session.wizard.set_hydration_error(error),
-        }
-
-        vec![self.launch_wizard_state_outbound()]
-    }
-
-    fn handle_issue_launch_wizard_prepared(
-        &mut self,
-        prepared: IssueLaunchWizardPrepared,
-    ) -> Vec<OutboundEvent> {
-        let IssueLaunchWizardPrepared {
-            client_id,
-            id,
-            knowledge_kind,
-            tab_id,
-            project_root,
-            issue_number,
-            result,
-        } = prepared;
-        if self.tab(&tab_id).is_none() {
-            return vec![OutboundEvent::reply(
-                &client_id,
-                BackendEvent::KnowledgeError {
-                    id,
-                    knowledge_kind,
-                    message: "Project tab not found".to_string(),
-                },
-            )];
-        }
-
-        match result {
-            Ok(branch_name) => match self.open_launch_wizard_for_branch(
-                &tab_id,
-                &project_root,
-                &branch_name,
-                Some(issue_number),
-            ) {
-                Ok(()) => vec![self.launch_wizard_state_outbound()],
-                Err(error) => vec![OutboundEvent::reply(
-                    &client_id,
-                    BackendEvent::KnowledgeError {
-                        id,
-                        knowledge_kind,
-                        message: error,
-                    },
-                )],
-            },
-            Err(error) => vec![OutboundEvent::reply(
-                &client_id,
-                BackendEvent::KnowledgeError {
-                    id,
-                    knowledge_kind,
-                    message: error,
-                },
-            )],
-        }
-    }
-
-    fn handle_launch_wizard_action(
-        &mut self,
-        action: gwt::LaunchWizardAction,
-        bounds: Option<WindowGeometry>,
-    ) -> Vec<OutboundEvent> {
-        let Some(mut session) = self.launch_wizard.take() else {
-            return Vec::new();
-        };
-        session.wizard.apply(action);
-
-        match session.wizard.completion.take() {
-            Some(LaunchWizardCompletion::Cancelled) => {
-                vec![self.launch_wizard_state_broadcast(None)]
-            }
-            Some(LaunchWizardCompletion::FocusWindow { window_id }) => {
-                let Some(address) = self.window_lookup.get(&window_id).cloned() else {
-                    session.wizard.error =
-                        Some("The selected session window is no longer available".to_string());
-                    self.launch_wizard = Some(session);
-                    return vec![self.launch_wizard_state_outbound()];
-                };
-                let Some(tab) = self.tab_mut(&address.tab_id) else {
-                    return Vec::new();
-                };
-                if !tab.workspace.focus_window(&address.raw_id, None) {
-                    session.wizard.error =
-                        Some("The selected session window is no longer available".to_string());
-                    self.launch_wizard = Some(session);
-                    return vec![self.launch_wizard_state_outbound()];
-                }
-                self.active_tab_id = Some(address.tab_id);
-                let _ = self.persist();
-                vec![
-                    self.workspace_state_broadcast(),
-                    self.launch_wizard_state_broadcast(None),
-                ]
-            }
-            Some(LaunchWizardCompletion::Launch(config)) => match *config {
-                LaunchWizardLaunchRequest::Agent(config) => {
-                    match self.spawn_agent_window(&session.tab_id, *config, bounds) {
-                        Ok(mut events) => {
-                            events.push(self.launch_wizard_state_broadcast(None));
-                            events
-                        }
-                        Err(error) => {
-                            session.wizard.error = Some(error);
-                            self.launch_wizard = Some(session);
-                            vec![self.launch_wizard_state_outbound()]
-                        }
-                    }
-                }
-                LaunchWizardLaunchRequest::Shell(config) => {
-                    match self.spawn_wizard_shell_window(&session.tab_id, *config, bounds) {
-                        Ok(mut events) => {
-                            events.push(self.launch_wizard_state_broadcast(None));
-                            events
-                        }
-                        Err(error) => {
-                            session.wizard.error = Some(error);
-                            self.launch_wizard = Some(session);
-                            vec![self.launch_wizard_state_outbound()]
-                        }
-                    }
-                }
-            },
-            None => {
-                self.launch_wizard = Some(session);
-                vec![self.launch_wizard_state_outbound()]
-            }
-        }
-    }
-
-    fn live_sessions_for_branch(&self, tab_id: &str, branch_name: &str) -> Vec<LiveSessionEntry> {
-        let mut entries = self
-            .active_agent_sessions
-            .values()
-            .filter(|session| session.tab_id == tab_id && session.branch_name == branch_name)
-            .map(|session| LiveSessionEntry {
-                session_id: session.session_id.clone(),
-                window_id: session.window_id.clone(),
-                agent_id: session.agent_id.clone(),
-                kind: "agent".to_string(),
-                name: session.display_name.clone(),
-                detail: Some(session.worktree_path.display().to_string()),
-                active: true,
-            })
-            .collect::<Vec<_>>();
-        entries.sort_by(|left, right| left.name.cmp(&right.name));
-        entries
-    }
-
-    fn active_session_branches_for_tab(&self, tab_id: &str) -> std::collections::HashSet<String> {
-        self.active_agent_sessions
-            .values()
-            .filter(|session| session.tab_id == tab_id)
-            .map(|session| session.branch_name.clone())
-            .collect()
-    }
-
-    fn handle_runtime_output(&mut self, id: String, data: Vec<u8>) -> Vec<OutboundEvent> {
-        if !self.window_lookup.contains_key(&id) {
-            return Vec::new();
-        }
-        vec![OutboundEvent::broadcast(BackendEvent::TerminalOutput {
-            id,
-            data_base64: base64::engine::general_purpose::STANDARD.encode(data),
-        })]
-    }
-
-    fn handle_runtime_status(
-        &mut self,
-        id: String,
-        status: WindowProcessStatus,
-        detail: Option<String>,
-    ) -> Vec<OutboundEvent> {
-        let should_auto_close =
-            should_auto_close_agent_window(&self.active_agent_sessions, &id, &status);
-        let Some(address) = self.window_lookup.get(&id).cloned() else {
-            self.deregister_pty_writer(&id);
-            self.runtimes.remove(&id);
-            self.window_details.remove(&id);
-            return Vec::new();
-        };
-
-        if let Some(tab) = self.tab_mut(&address.tab_id) {
-            let _ = tab.workspace.set_status(&address.raw_id, status.clone());
-        }
-        match detail.as_ref() {
-            Some(detail) if !detail.is_empty() => {
-                self.window_details.insert(id.clone(), detail.clone());
-            }
-            _ => {
-                self.window_details.remove(&id);
-            }
-        }
-        if should_auto_close {
-            self.stop_window_runtime(&id);
-            if !close_window_from_workspace(
-                &mut self.tabs,
-                &mut self.window_lookup,
-                &mut self.window_details,
-                &id,
-            ) {
-                return Vec::new();
-            }
-            let _ = self.persist();
-            return vec![self.workspace_state_broadcast()];
-        }
-        if matches!(
-            status,
-            WindowProcessStatus::Error | WindowProcessStatus::Exited
-        ) {
-            self.runtimes.remove(&id);
-            self.mark_agent_session_stopped(&id);
-        }
-        let _ = self.persist();
-
-        vec![
-            self.workspace_state_broadcast(),
-            OutboundEvent::broadcast(BackendEvent::TerminalStatus { id, status, detail }),
-        ]
-    }
-
-    fn handle_launch_complete(
-        &mut self,
-        window_id: String,
-        result: Result<
-            (
-                ProcessLaunch,
-                String,
-                String,
-                String,
-                PathBuf,
-                gwt_agent::AgentId,
-            ),
-            String,
-        >,
-    ) -> Vec<OutboundEvent> {
-        match result {
-            Ok((
-                process_launch,
-                session_id,
-                branch_name,
-                display_name,
-                worktree_path,
-                agent_id,
-            )) => {
-                let Some(address) = self.window_lookup.get(&window_id).cloned() else {
-                    return vec![OutboundEvent::broadcast(BackendEvent::TerminalStatus {
-                        id: window_id,
-                        status: WindowProcessStatus::Error,
-                        detail: Some("Window not found".to_string()),
-                    })];
-                };
-                let Some(tab) = self.tab(&address.tab_id) else {
-                    return vec![OutboundEvent::broadcast(BackendEvent::TerminalStatus {
-                        id: window_id,
-                        status: WindowProcessStatus::Error,
-                        detail: Some("Project tab not found".to_string()),
-                    })];
-                };
-                let Some(window) = tab.workspace.window(&address.raw_id) else {
-                    return vec![OutboundEvent::broadcast(BackendEvent::TerminalStatus {
-                        id: window_id,
-                        status: WindowProcessStatus::Error,
-                        detail: Some("Window not found".to_string()),
-                    })];
-                };
-                let geometry = window.geometry.clone();
-
-                self.active_agent_sessions.insert(
-                    window_id.clone(),
-                    ActiveAgentSession {
-                        window_id: window_id.clone(),
-                        session_id,
-                        agent_id: agent_id.to_string(),
-                        branch_name,
-                        display_name,
-                        worktree_path,
-                        tab_id: address.tab_id,
-                    },
-                );
-
-                let _ = self.persist();
-
-                match self.spawn_process_window(&window_id, geometry, process_launch) {
-                    Ok(event) => vec![
-                        self.workspace_state_broadcast(),
-                        OutboundEvent::broadcast(event),
-                    ],
-                    Err(error) => vec![OutboundEvent::broadcast(BackendEvent::TerminalStatus {
-                        id: window_id,
-                        status: WindowProcessStatus::Error,
-                        detail: Some(error),
-                    })],
-                }
-            }
-            Err(error) => vec![OutboundEvent::broadcast(BackendEvent::TerminalStatus {
-                id: window_id,
-                status: WindowProcessStatus::Error,
-                detail: Some(error),
-            })],
-        }
-    }
-
-    fn handle_shell_launch_complete(
-        &mut self,
-        window_id: String,
-        result: Result<ProcessLaunch, String>,
-    ) -> Vec<OutboundEvent> {
-        match result {
-            Ok(process_launch) => {
-                let Some(address) = self.window_lookup.get(&window_id).cloned() else {
-                    return vec![OutboundEvent::broadcast(BackendEvent::TerminalStatus {
-                        id: window_id,
-                        status: WindowProcessStatus::Error,
-                        detail: Some("Window not found".to_string()),
-                    })];
-                };
-                let Some(tab) = self.tab(&address.tab_id) else {
-                    return vec![OutboundEvent::broadcast(BackendEvent::TerminalStatus {
-                        id: window_id,
-                        status: WindowProcessStatus::Error,
-                        detail: Some("Project tab not found".to_string()),
-                    })];
-                };
-                let Some(window) = tab.workspace.window(&address.raw_id) else {
-                    return vec![OutboundEvent::broadcast(BackendEvent::TerminalStatus {
-                        id: window_id,
-                        status: WindowProcessStatus::Error,
-                        detail: Some("Window not found".to_string()),
-                    })];
-                };
-                let geometry = window.geometry.clone();
-
-                match self.spawn_process_window(&window_id, geometry, process_launch) {
-                    Ok(event) => vec![
-                        self.workspace_state_broadcast(),
-                        OutboundEvent::broadcast(event),
-                    ],
-                    Err(error) => vec![OutboundEvent::broadcast(BackendEvent::TerminalStatus {
-                        id: window_id,
-                        status: WindowProcessStatus::Error,
-                        detail: Some(error),
-                    })],
-                }
-            }
-            Err(error) => vec![OutboundEvent::broadcast(BackendEvent::TerminalStatus {
-                id: window_id,
-                status: WindowProcessStatus::Error,
-                detail: Some(error),
-            })],
-        }
-    }
-
-    fn start_window(
-        &mut self,
-        tab_id: &str,
-        raw_id: &str,
-        preset: WindowPreset,
-        geometry: WindowGeometry,
-    ) -> Option<BackendEvent> {
-        self.register_window(tab_id, raw_id);
-        let window_id = combined_window_id(tab_id, raw_id);
-        if !preset.requires_process() {
-            self.set_window_status(tab_id, raw_id, WindowProcessStatus::Ready);
-            return None;
-        }
-
-        let project_root = self
-            .tab(tab_id)
-            .map(|tab| tab.project_root.clone())
-            .unwrap_or_else(|| PathBuf::from("."));
-
-        let shell = match detect_shell_program() {
-            Ok(shell) => shell,
-            Err(error) => {
-                self.set_window_status(tab_id, raw_id, WindowProcessStatus::Error);
-                self.window_details
-                    .insert(window_id.clone(), error.to_string());
-                return Some(BackendEvent::TerminalStatus {
-                    id: window_id,
-                    status: WindowProcessStatus::Error,
-                    detail: Some(error.to_string()),
-                });
-            }
-        };
-
-        let launch = match resolve_launch_spec_with_fallback(preset, &shell) {
-            Ok(launch) => launch,
-            Err(error) => {
-                self.set_window_status(tab_id, raw_id, WindowProcessStatus::Error);
-                self.window_details
-                    .insert(window_id.clone(), error.to_string());
-                return Some(BackendEvent::TerminalStatus {
-                    id: window_id,
-                    status: WindowProcessStatus::Error,
-                    detail: Some(error.to_string()),
-                });
-            }
-        };
-
-        match self.spawn_process_window(
-            &window_id,
-            geometry,
-            ProcessLaunch {
-                command: launch.command,
-                args: launch.args,
-                env: spawn_env(),
-                cwd: Some(project_root),
-            },
-        ) {
-            Ok(event) => Some(event),
-            Err(error) => {
-                self.set_window_status(tab_id, raw_id, WindowProcessStatus::Error);
-                self.window_details.insert(window_id.clone(), error.clone());
-                Some(BackendEvent::TerminalStatus {
-                    id: window_id,
-                    status: WindowProcessStatus::Error,
-                    detail: Some(error),
-                })
-            }
-        }
-    }
-
-    fn spawn_process_window(
-        &mut self,
-        id: &str,
-        geometry: WindowGeometry,
-        launch: ProcessLaunch,
-    ) -> Result<BackendEvent, String> {
-        let (cols, rows) = geometry_to_pty_size(&geometry);
-        let pane = Pane::new(
-            id.to_string(),
-            launch.command,
-            launch.args,
-            cols,
-            rows,
-            launch.env,
-            launch.cwd,
-        )
-        .map_err(|error| error.to_string())?;
-        let pane = Arc::new(Mutex::new(pane));
-
-        let output_thread = self.spawn_output_thread(id.to_string(), pane.clone());
-        if let Some(address) = self.window_lookup.get(id).cloned() {
-            self.set_window_status(
-                &address.tab_id,
-                &address.raw_id,
-                WindowProcessStatus::Running,
-            );
-        }
-        self.window_details.remove(id);
-        // Publish the PTY handle to the WebSocket fast-path registry BEFORE
-        // inserting the runtime so that the first `terminal_input` from the
-        // frontend (which can arrive immediately after `TerminalStatus`) has a
-        // target to write to. Registry holds a cloned `Arc<PtyHandle>`; the
-        // real owner remains the `Mutex<Pane>` in `WindowRuntime`.
-        self.register_pty_writer(id, &pane);
-        self.runtimes.insert(
-            id.to_string(),
-            WindowRuntime {
-                pane,
-                output_thread: Some(output_thread),
-            },
-        );
-        Ok(BackendEvent::TerminalStatus {
-            id: id.to_string(),
-            status: WindowProcessStatus::Running,
-            detail: None,
-        })
-    }
-
-    fn spawn_agent_window(
-        &mut self,
-        tab_id: &str,
-        config: gwt_agent::LaunchConfig,
-        bounds: Option<WindowGeometry>,
-    ) -> Result<Vec<OutboundEvent>, String> {
-        let tab = self
-            .tab_mut(tab_id)
-            .ok_or_else(|| "Project tab not found".to_string())?;
-        let project_root = tab.project_root.display().to_string();
-        let title = format!(
-            "{} · {}",
-            config.display_name,
-            config.branch.as_ref().unwrap_or(&"workspace".to_string())
-        );
-        let default_bounds = WindowGeometry {
-            x: 100.0,
-            y: 40.0,
-            width: 1000.0,
-            height: 760.0,
-        };
-        let window = tab.workspace.add_window_with_title(
-            WindowPreset::Agent,
-            title.clone(),
-            false,
-            bounds.unwrap_or(default_bounds),
-        );
-        self.register_window(tab_id, &window.id);
-        let window_id = combined_window_id(tab_id, &window.id);
-
-        let events = vec![
-            self.workspace_state_broadcast(),
-            OutboundEvent::broadcast(BackendEvent::TerminalStatus {
-                id: window_id.clone(),
-                status: WindowProcessStatus::Starting,
-                detail: None,
-            }),
-        ];
-
-        let proxy = self.proxy.clone();
-        let sessions_dir = self.sessions_dir.clone();
-        let hook_forward_target = self.hook_forward_target.clone();
-
-        thread::spawn(move || {
-            Self::spawn_agent_window_async(
-                proxy,
-                sessions_dir,
-                project_root,
-                window_id,
-                config,
-                hook_forward_target,
-            )
-        });
-
-        Ok(events)
-    }
-
-    fn spawn_wizard_shell_window(
-        &mut self,
-        tab_id: &str,
-        config: ShellLaunchConfig,
-        bounds: Option<WindowGeometry>,
-    ) -> Result<Vec<OutboundEvent>, String> {
-        let tab = self
-            .tab_mut(tab_id)
-            .ok_or_else(|| "Project tab not found".to_string())?;
-        let project_root = tab.project_root.display().to_string();
-        let title = format!(
-            "{} · {}",
-            config.display_name,
-            config.branch.as_ref().unwrap_or(&"workspace".to_string())
-        );
-        let default_bounds = WindowGeometry {
-            x: 100.0,
-            y: 40.0,
-            width: 1000.0,
-            height: 760.0,
-        };
-        let window = tab.workspace.add_window_with_title(
-            WindowPreset::Shell,
-            title,
-            false,
-            bounds.unwrap_or(default_bounds),
-        );
-        self.register_window(tab_id, &window.id);
-        let window_id = combined_window_id(tab_id, &window.id);
-
-        let events = vec![
-            self.workspace_state_broadcast(),
-            OutboundEvent::broadcast(BackendEvent::TerminalStatus {
-                id: window_id.clone(),
-                status: WindowProcessStatus::Starting,
-                detail: None,
-            }),
-        ];
-
-        let proxy = self.proxy.clone();
-        thread::spawn(move || {
-            Self::spawn_wizard_shell_window_async(proxy, project_root, window_id, config)
-        });
-
-        Ok(events)
-    }
-
-    fn spawn_agent_window_async(
-        proxy: AppEventProxy,
-        sessions_dir: PathBuf,
-        project_root: String,
-        window_id: String,
-        mut config: gwt_agent::LaunchConfig,
-        hook_forward_target: Option<HookForwardTarget>,
-    ) {
-        let result = (|| {
-            proxy.send(UserEvent::LaunchProgress {
-                window_id: window_id.clone(),
-                message: "Preparing worktree...".to_string(),
-            });
-            resolve_launch_worktree(Path::new(&project_root), &mut config)?;
-
-            proxy.send(UserEvent::LaunchProgress {
-                window_id: window_id.clone(),
-                message: "Starting Docker service...".to_string(),
-            });
-            apply_docker_runtime_to_launch_config(Path::new(&project_root), &mut config)?;
-
-            proxy.send(UserEvent::LaunchProgress {
-                window_id: window_id.clone(),
-                message: "Configuring workspace...".to_string(),
-            });
-            let worktree_path = config
-                .working_dir
-                .clone()
-                .unwrap_or_else(|| PathBuf::from(&project_root));
-            refresh_managed_gwt_assets_for_worktree(&worktree_path)
-                .map_err(|error| error.to_string())?;
-            if let Err(error) = gwt::index_worker::bootstrap_project_index_for_path(&worktree_path)
-            {
-                tracing::warn!(
-                    worktree = %worktree_path.display(),
-                    error = %error,
-                    "project index bootstrap skipped during worktree prepare"
-                );
-            }
-
-            if config.runtime_target == gwt_agent::LaunchRuntimeTarget::Host
-                && apply_host_package_runner_fallback(&mut config)
-            {
-                proxy.send(UserEvent::LaunchProgress {
-                    window_id: window_id.clone(),
-                    message: "bunx unavailable, switching to npx...".to_string(),
-                });
-            }
-            install_launch_gwt_bin_env(&mut config.env_vars, config.runtime_target)?;
-
-            let branch_name = config
-                .branch
-                .clone()
-                .unwrap_or_else(|| "workspace".to_string());
-
-            let agent_id = config.agent_id.clone();
-            let mut session =
-                gwt_agent::Session::new(&worktree_path, branch_name.clone(), agent_id.clone());
-            session.display_name = config.display_name.clone();
-            session.tool_version = config.tool_version.clone();
-            session.model = config.model.clone();
-            session.reasoning_level = config.reasoning_level.clone();
-            session.skip_permissions = config.skip_permissions;
-            session.codex_fast_mode = config.codex_fast_mode;
-            session.runtime_target = config.runtime_target;
-            session.docker_service = config.docker_service.clone();
-            session.docker_lifecycle_intent = config.docker_lifecycle_intent;
-            session.linked_issue_number = config.linked_issue_number;
-            session.launch_command = config.command.clone();
-            session.launch_args = config.args.clone();
-            session.update_status(gwt_agent::AgentStatus::Running);
-
-            let session_id = session.id.clone();
-            let runtime_path = gwt_agent::runtime_state_path(&sessions_dir, &session_id);
-            config.env_vars.insert(
-                gwt_agent::GWT_SESSION_ID_ENV.to_string(),
-                session_id.clone(),
-            );
-            config.env_vars.insert(
-                gwt_agent::GWT_SESSION_RUNTIME_PATH_ENV.to_string(),
-                runtime_path.display().to_string(),
-            );
-            if let Some(target) = hook_forward_target {
-                config
-                    .env_vars
-                    .insert(gwt_agent::GWT_HOOK_FORWARD_URL_ENV.to_string(), target.url);
-                config.env_vars.insert(
-                    gwt_agent::GWT_HOOK_FORWARD_TOKEN_ENV.to_string(),
-                    target.token,
-                );
-            }
-            config
-                .env_vars
-                .entry("COLORTERM".to_string())
-                .or_insert_with(|| "truecolor".to_string());
-            finalize_docker_agent_launch_config(Path::new(&project_root), &mut config)?;
-
-            session
-                .save(&sessions_dir)
-                .map_err(|error| error.to_string())?;
-            gwt_agent::SessionRuntimeState::new(gwt_agent::AgentStatus::Running)
-                .save(&runtime_path)
-                .map_err(|error| error.to_string())?;
-
-            let process_launch = ProcessLaunch {
-                command: config.command.clone(),
-                args: config.args.clone(),
-                env: config.env_vars.clone(),
-                cwd: config.working_dir.clone(),
-            };
-
-            Ok((
-                process_launch,
-                session_id,
-                branch_name,
-                config.display_name,
-                worktree_path,
-                agent_id,
-            ))
-        })();
-
-        match result {
-            Ok((
-                process_launch,
-                session_id,
-                branch_name,
-                display_name,
-                worktree_path,
-                agent_id,
-            )) => {
-                proxy.send(UserEvent::LaunchComplete {
-                    window_id,
-                    result: Ok((
-                        process_launch,
-                        session_id,
-                        branch_name,
-                        display_name,
-                        worktree_path,
-                        agent_id,
-                    )),
-                });
-            }
-            Err(error) => {
-                proxy.send(UserEvent::LaunchComplete {
-                    window_id,
-                    result: Err(error),
-                });
-            }
-        }
-    }
-
-    fn spawn_wizard_shell_window_async(
-        proxy: AppEventProxy,
-        project_root: String,
-        window_id: String,
-        mut config: ShellLaunchConfig,
-    ) {
-        let result = (|| {
-            proxy.send(UserEvent::LaunchProgress {
-                window_id: window_id.clone(),
-                message: "Preparing worktree...".to_string(),
-            });
-            resolve_shell_launch_worktree(Path::new(&project_root), &mut config)?;
-
-            if config.runtime_target == gwt_agent::LaunchRuntimeTarget::Docker {
-                proxy.send(UserEvent::LaunchProgress {
-                    window_id: window_id.clone(),
-                    message: "Starting Docker service...".to_string(),
-                });
-            }
-
-            build_shell_process_launch(Path::new(&project_root), &mut config)
-        })();
-
-        proxy.send(UserEvent::ShellLaunchComplete { window_id, result });
-    }
-
-    fn mark_agent_session_stopped(&mut self, window_id: &str) {
-        let Some(session) = self.active_agent_sessions.remove(window_id) else {
-            return;
-        };
-        let _ = gwt_agent::persist_session_status(
-            &self.sessions_dir,
-            &session.session_id,
-            gwt_agent::AgentStatus::Stopped,
-        );
-    }
-
-    fn register_pty_writer(&self, id: &str, pane: &Arc<Mutex<Pane>>) {
-        let Ok(pane_guard) = pane.lock() else {
-            tracing::warn!(
-                target: "gwt_input_trace",
-                stage = "registry_lock_poisoned",
-                window_id = %id,
-                "failed to register PTY writer: pane mutex poisoned"
-            );
-            return;
-        };
-        let pty = pane_guard.shared_pty();
-        drop(pane_guard);
-        match self.pty_writers.write() {
-            Ok(mut guard) => {
-                guard.insert(id.to_string(), pty);
-            }
-            Err(error) => {
-                tracing::warn!(
-                    target: "gwt_input_trace",
-                    stage = "registry_write_poisoned",
-                    window_id = %id,
-                    error = %error,
-                    "failed to register PTY writer: registry poisoned"
-                );
-            }
-        }
-    }
-
-    fn deregister_pty_writer(&self, id: &str) {
-        match self.pty_writers.write() {
-            Ok(mut guard) => {
-                guard.remove(id);
-            }
-            Err(error) => {
-                tracing::warn!(
-                    target: "gwt_input_trace",
-                    stage = "registry_deregister_poisoned",
-                    window_id = %id,
-                    error = %error,
-                    "failed to deregister PTY writer: registry poisoned"
-                );
-            }
-        }
-    }
-
-    fn stop_window_runtime(&mut self, window_id: &str) {
-        self.mark_agent_session_stopped(window_id);
-        self.deregister_pty_writer(window_id);
-        if let Some(mut runtime) = self.runtimes.remove(window_id) {
-            if let Ok(pane) = runtime.pane.lock() {
-                let _ = pane.kill();
-            }
-            if let Some(handle) = runtime.output_thread.take() {
-                // PTY and its process group were already terminated by
-                // `pane.kill()`, so the reader should see EOF quickly. Cap
-                // the wait anyway so shutdown never stalls the event loop
-                // if a stuck syscall keeps the reader in `read`. If the
-                // timeout elapses the reader thread is detached; its Arc
-                // clone of the Pane will still be released when the thread
-                // does finally observe EOF.
-                let (tx, rx) = std_mpsc::channel();
-                thread::spawn(move || {
-                    let _ = handle.join();
-                    let _ = tx.send(());
-                });
-                let _ = rx.recv_timeout(Duration::from_millis(500));
-            }
-        }
-        self.window_details.remove(window_id);
-    }
-
-    /// Stop every active window runtime. Called from the application shutdown
-    /// paths so no PTY / agent process outlives the GUI.
-    fn stop_all_runtimes(&mut self) {
-        let ids: Vec<String> = self.runtimes.keys().cloned().collect();
-        for id in ids {
-            self.stop_window_runtime(&id);
-        }
-    }
-
-    fn spawn_output_thread(&self, id: String, pane: Arc<Mutex<Pane>>) -> JoinHandle<()> {
-        let proxy = self.proxy.clone();
-        thread::spawn(move || {
-            let reader = match pane
-                .lock()
-                .map_err(|error| error.to_string())
-                .and_then(|pane| pane.reader().map_err(|error| error.to_string()))
-            {
-                Ok(reader) => reader,
-                Err(error) => {
-                    proxy.send(UserEvent::RuntimeStatus {
-                        id,
-                        status: WindowProcessStatus::Error,
-                        detail: Some(error),
-                    });
-                    return;
-                }
-            };
-
-            let mut reader = reader;
-            let mut buffer = [0u8; 4096];
-            loop {
-                match reader.read(&mut buffer) {
-                    Ok(0) => break,
-                    Ok(read) => {
-                        let chunk = buffer[..read].to_vec();
-                        let lock_started = Instant::now();
-                        if let Ok(mut pane) = pane.lock() {
-                            let lock_wait_us = lock_started.elapsed().as_micros() as u64;
-                            let parse_started = Instant::now();
-                            pane.process_bytes(&chunk);
-                            let parse_us = parse_started.elapsed().as_micros() as u64;
-                            // Log only when the contention window is large enough
-                            // to plausibly starve a concurrent `write_input`. The
-                            // threshold keeps the log volume bounded during
-                            // normal output bursts while still surfacing the
-                            // lock-hold windows that matter for drop triage.
-                            if lock_wait_us > 500 || parse_us > 500 {
-                                tracing::debug!(
-                                    target: "gwt_input_trace",
-                                    stage = "reader_pane_lock",
-                                    window_id = %id,
-                                    chunk_len = read,
-                                    lock_wait_us,
-                                    parse_us,
-                                    "reader thread held pane mutex (output parsing)"
-                                );
-                            }
-                        }
-                        proxy.send(UserEvent::RuntimeOutput {
-                            id: id.clone(),
-                            data: chunk,
-                        });
-                    }
-                    Err(error) => {
-                        proxy.send(UserEvent::RuntimeStatus {
-                            id: id.clone(),
-                            status: WindowProcessStatus::Error,
-                            detail: Some(error.to_string()),
-                        });
-                        return;
-                    }
-                }
-            }
-
-            let status = pane
-                .lock()
-                .map_err(|error| error.to_string())
-                .and_then(|mut pane| {
-                    pane.check_status()
-                        .cloned()
-                        .map_err(|error| error.to_string())
-                });
-
-            match status {
-                Ok(PaneStatus::Running) | Ok(PaneStatus::Completed(0)) => {
-                    proxy.send(UserEvent::RuntimeStatus {
-                        id,
-                        status: WindowProcessStatus::Exited,
-                        detail: Some("Process exited".to_string()),
-                    });
-                }
-                Ok(PaneStatus::Completed(code)) => {
-                    proxy.send(UserEvent::RuntimeStatus {
-                        id,
-                        status: WindowProcessStatus::Error,
-                        detail: Some(format!("Process exited with status {code}")),
-                    });
-                }
-                Ok(PaneStatus::Error(message)) => {
-                    proxy.send(UserEvent::RuntimeStatus {
-                        id,
-                        status: WindowProcessStatus::Error,
-                        detail: Some(message),
-                    });
-                }
-                Err(error) => {
-                    proxy.send(UserEvent::RuntimeStatus {
-                        id,
-                        status: WindowProcessStatus::Error,
-                        detail: Some(error),
-                    });
-                }
-            }
-        })
-    }
-
-    fn app_state_view(&self) -> gwt::AppStateView {
-        app_state_view_from_parts(
-            &self.tabs,
-            self.active_tab_id.as_deref(),
-            &self.recent_projects,
-        )
-    }
-
-    fn workspace_state_broadcast(&self) -> OutboundEvent {
-        OutboundEvent::broadcast(BackendEvent::WorkspaceState {
-            workspace: self.app_state_view(),
-        })
-    }
-
-    fn launch_wizard_state_outbound(&self) -> OutboundEvent {
-        OutboundEvent::broadcast(BackendEvent::LaunchWizardState {
-            wizard: self
-                .launch_wizard
-                .as_ref()
-                .map(|wizard| Box::new(wizard.wizard.view())),
-        })
-    }
-
-    fn launch_wizard_state_broadcast(
-        &self,
-        wizard: Option<gwt::LaunchWizardView>,
-    ) -> OutboundEvent {
-        OutboundEvent::broadcast(BackendEvent::LaunchWizardState {
-            wizard: wizard.map(Box::new),
-        })
-    }
-
-    fn window_status(&self, window_id: &str) -> Option<WindowProcessStatus> {
-        let address = self.window_lookup.get(window_id)?;
-        let tab = self.tab(&address.tab_id)?;
-        let window = tab.workspace.window(&address.raw_id)?;
-        Some(window.status.clone())
-    }
-
-    fn register_window(&mut self, tab_id: &str, raw_id: &str) {
-        self.window_lookup.insert(
-            combined_window_id(tab_id, raw_id),
-            WindowAddress {
-                tab_id: tab_id.to_string(),
-                raw_id: raw_id.to_string(),
-            },
-        );
-    }
-
-    fn set_window_status(&mut self, tab_id: &str, raw_id: &str, status: WindowProcessStatus) {
-        if let Some(tab) = self.tab_mut(tab_id) {
-            let _ = tab.workspace.set_status(raw_id, status);
-        }
-    }
-
-    fn resize_runtime_to_window(&self, window_id: &str) {
-        let Some(address) = self.window_lookup.get(window_id) else {
-            return;
-        };
-        let Some(tab) = self.tab(&address.tab_id) else {
-            return;
-        };
-        let Some(window) = tab.workspace.window(&address.raw_id) else {
-            return;
-        };
-        if !window.preset.requires_process() {
-            return;
-        }
-        if let Some(runtime) = self.runtimes.get(window_id) {
-            if let Ok(mut pane) = runtime.pane.lock() {
-                let (cols, rows) = geometry_to_pty_size(&window.geometry);
-                let _ = pane.resize(cols.max(20), rows.max(6));
-            }
-        }
-    }
-
-    fn tab(&self, tab_id: &str) -> Option<&ProjectTabRuntime> {
-        self.tabs.iter().find(|tab| tab.id == tab_id)
-    }
-
-    fn tab_mut(&mut self, tab_id: &str) -> Option<&mut ProjectTabRuntime> {
-        self.tabs.iter_mut().find(|tab| tab.id == tab_id)
-    }
-
-    fn active_tab_mut(&mut self) -> Option<&mut ProjectTabRuntime> {
-        let active_tab_id = self.active_tab_id.clone()?;
-        self.tab_mut(&active_tab_id)
-    }
-
-    fn set_active_tab(&mut self, tab_id: String) -> bool {
-        let wizard_closed = self
-            .launch_wizard
-            .as_ref()
-            .is_some_and(|wizard| wizard.tab_id != tab_id);
-        self.active_tab_id = Some(tab_id);
-        if wizard_closed {
-            self.launch_wizard = None;
-        }
-        wizard_closed
-    }
-
-    fn clear_launch_wizard(&mut self) -> Option<LaunchWizardSession> {
-        self.launch_wizard.take()
-    }
-
-    fn rebuild_window_lookup(&mut self) {
-        self.window_lookup.clear();
-        let pairs = self
-            .tabs
-            .iter()
-            .flat_map(|tab| {
-                tab.workspace
-                    .persisted()
-                    .windows
-                    .iter()
-                    .map(|window| (tab.id.clone(), window.id.clone()))
-                    .collect::<Vec<_>>()
-            })
-            .collect::<Vec<_>>();
-        for (tab_id, raw_id) in pairs {
-            self.register_window(&tab_id, &raw_id);
-        }
-    }
-
-    fn seed_restored_window_details(&mut self) {
-        for tab in &self.tabs {
-            for window in &tab.workspace.persisted().windows {
-                if window.preset.requires_process() && window.status == WindowProcessStatus::Exited
-                {
-                    self.window_details.insert(
-                        combined_window_id(&tab.id, &window.id),
-                        "Restored window is paused. Launch a new terminal when you want to start it."
-                            .to_string(),
-                    );
-                }
-            }
-        }
-    }
-
-    fn persist(&self) -> std::io::Result<()> {
-        save_session_state(
-            &self.session_state_path,
-            &gwt::PersistedSessionState {
-                tabs: self
-                    .tabs
-                    .iter()
-                    .map(|tab| gwt::PersistedSessionTabState {
-                        id: tab.id.clone(),
-                        title: tab.title.clone(),
-                        project_root: tab.project_root.clone(),
-                        kind: tab.kind,
-                    })
-                    .collect(),
-                active_tab_id: normalize_active_tab_id(&self.tabs, self.active_tab_id.clone()),
-                recent_projects: self.recent_projects.clone(),
-            },
-        )?;
-
-        for tab in &self.tabs {
-            save_workspace_state(
-                &workspace_state_path(&tab.project_root),
-                &tab.workspace.persistable_state(),
-            )?;
-        }
-
-        Ok(())
-    }
 }
 
 fn combined_window_id(tab_id: &str, raw_id: &str) -> String {
